Create lineage.test.js (closes #120)
